chore(#23): lint 에러 수정
