+++ src/index.tsx
@@ -3,10 +3,10 @@ import { StrictMode } from "react";
 import ReactDOM from "react-dom/client";
 import { routeTree } from "@/shared/lib/@generated/routeTree.gen";
 import "@/shared/lib/globals.css";
+import { ErrorCatcher } from "./shared/ui/error-boundary/error-catcher";
+import QueryClientBoundary from "./shared/ui/error-boundary/query-client-boundary";
 import UnPredictableErrorBoundary from "./shared/ui/error-boundary/unpredictable-error-boundary";
 import ToastContainer from "./shared/ui/toast/toast-container";
-import QueryClientBoundary from "./shared/ui/error-boundary/query-client-boundary";
-import { ErrorCatcher } from "./shared/ui/error-boundary/error-catcher";
 
 const router = createRouter({
 	routeTree,
@@ -18,8 +18,8 @@ declare module "@tanstack/react-router" {
 	}
 }
 
-const rootElement = document.getElementById("root")!;
-if (!rootElement.innerHTML) {
+const rootElement = document.getElementById("root");
+if (rootElement && !rootElement.innerHTML) {
 	const root = ReactDOM.createRoot(rootElement);
 	root.render(
 		<StrictMode>
