--- src/shared/lib/is-request-error.ts
@@ -1,7 +1,0 @@
-import RequestError from "../ui/error-boundary/request-error";
-
-// const isRequestError = (error: Error) => {
-// 	return error instanceof RequestError;
-// };
-
-// export default isRequestError;
