+++ src/shared/lib/query-client.ts
@@ -1,9 +1,6 @@
 import { MutationCache, QueryCache, QueryClient } from "@tanstack/react-query";
-import type { Method } from "axios";
 import type { ApiErrorResponse } from "@/shared/api";
 import { useAppErrorStore } from "../ui/error-boundary/app-error-store";
-
-import { toast } from "./toast";
 import { isApiError, isAuthError, isServerError } from "./error-type-fn";
 
 type ErrorHandlingStrategy = "toast" | "errorBoundary" | "ignore";
