+++ src/shared/lib/toast/event-manager.ts
@@ -1,7 +1,10 @@
 import type { AddEventHandlerArgs, ToastEventCallbackMap } from "./types";
 
 export const toastEventManager = (() => {
-	const eventCallbackList = new Map<keyof ToastEventCallbackMap, any>();
+	const eventCallbackList = new Map<
+		keyof ToastEventCallbackMap,
+		ToastEventCallbackMap[keyof ToastEventCallbackMap]
+	>();
 
 	return {
 		addEventHandler<K extends keyof ToastEventCallbackMap>({
@@ -11,7 +14,10 @@ export const toastEventManager = (() => {
 			eventCallbackList.set(eventType, callback);
 		},
 
-		trigger<K extends keyof ToastEventCallbackMap>(eventType: K, ...args: any) {
+		trigger<K extends keyof ToastEventCallbackMap>(
+			eventType: K,
+			...args: unknown[]
+		) {
 			if (!eventCallbackList.has(eventType)) {
 				throw new Error(
 					`토스트 이벤트 핸들러에 등록된 ${eventType} 이벤트가 없습니다. 이벤트 등록 후 호출해주세요.`,
@@ -19,7 +25,9 @@ export const toastEventManager = (() => {
 			}
 
 			const callback = eventCallbackList.get(eventType);
-			callback(...args);
+			if (typeof callback === "function") {
+				(callback as (...a: unknown[]) => void)(...args);
+			}
 		},
 	};
 })();
