+++ src/shared/lib/toast/index.ts
@@ -1,18 +1,16 @@
-export { TOAST_EVENT } from "./types";
+export { default as toast } from "@/shared/lib/toast/api";
 export { toastEventManager } from "./event-manager";
-export { useToast } from "./use-toast";
-
 // Toast 타입 (UI 컴포넌트와 공유)
 export type {
-	ToastPosition,
-	ToastOptions,
-	ToastMessage,
-	ToastArgs,
-	ToastType,
-	ToastProps,
-	ToastEventType,
-	ToastEventCallbackMap,
 	AddEventHandlerArgs,
+	ToastArgs,
+	ToastEventCallbackMap,
+	ToastEventType,
+	ToastMessage,
+	ToastOptions,
+	ToastPosition,
+	ToastProps,
+	ToastType,
 } from "./types";
-
-export { default as toast } from "@/shared/lib/toast/api";
+export { TOAST_EVENT } from "./types";
+export { useToast } from "./use-toast";
