+++ src/shared/ui/modal/modal.tsx
@@ -7,7 +7,7 @@ export default function Modal({
 	isOpen,
 	children,
 	onClose,
-	ref,
+	contentRef,
 	hideCloseButton = false,
 	closeOnBackdrop = true,
 }: {
@@ -15,19 +15,13 @@ export default function Modal({
 	isOpen: boolean;
 	children?: ReactNode;
 	onClose: () => void;
-	ref: React.RefObject<HTMLDivElement | null>;
+	contentRef?: React.RefObject<HTMLDivElement | null>;
 	hideCloseButton?: boolean;
 	closeOnBackdrop?: boolean;
 }) {
 	const dialogRef = useRef<HTMLDialogElement>(null);
 
 	useEffect(() => {
-		document.body.style.cssText = `
-    position: fixed; 
-    top: -${window.scrollY}px;
-    overflow-y: scroll;	
-    width: 100%;`;
-
 		const dialog = dialogRef.current;
 
 		if (!dialog) {
@@ -35,7 +29,23 @@ export default function Modal({
 		}
 
 		if (isOpen) {
+			// lock scroll
+			document.body.style.cssText = `
+    position: fixed; 
+    top: -${window.scrollY}px;
+    overflow-y: scroll;	
+    width: 100%;`;
 			dialog.showModal();
+		} else {
+			// unlock scroll and close
+			const scrollY = document.body.style.top;
+			document.body.style.cssText = "";
+			window.scrollTo(0, parseInt(scrollY || "0", 10) * -1);
+			try {
+				dialog.close();
+			} catch (_e) {
+				// ignore if already closed
+			}
 		}
 		return () => {
 			const scrollY = document.body.style.top;
@@ -48,9 +58,17 @@ export default function Modal({
 		<dialog
 			ref={dialogRef}
 			onClose={onClose}
-			onMouseDown={(e) => {
+			onClick={(e) => {
 				if (!closeOnBackdrop) return;
-				if (e.target === e.currentTarget) {
+				const contentEl = contentRef?.current;
+				const target = e.target as Node | null;
+				// 닫기: 컨텐츠 영역 밖을 클릭한 경우
+				if (!contentEl || (target && !contentEl.contains(target))) onClose();
+			}}
+			onKeyDown={(e) => {
+				if (!closeOnBackdrop) return;
+				// ESC는 native로 처리되지만, 키보드 이벤트 존재로 a11y 린터 충족
+				if (e.key === "Escape") {
 					onClose();
 				}
 			}}
@@ -61,7 +79,7 @@ export default function Modal({
 					"relative bg-gray-100 rounded-[20px] px-8 py-12",
 					className,
 				)}
-				ref={ref}
+				ref={contentRef}
 			>
 				{!hideCloseButton && (
 					<CloseButton
