+++ src/widgets/notification/lib/types.ts
